Remove unneeded or invisible SVG elements from official logos
The logos had some strange elements which seemed to be intended as a shadow for the inside of the last "e" in the "free software" text, however in most/all of the logos they were misplaced and/or the shape did not fit the letter "e". I decided that it's best for consistency if I removed these elements. (Most of the logos look exactly the same when rendered, and the ones that *have* changed were only changed a little bit - you most likely wouldn't notice the change unless you were actively looking for it)

Signed-off-by: Дима Ш. <dima.o.sh@proton.me>